+++ star-rx72n-firmware/libs/rx_hal/src/rx_tmr.c
@@ -219,10 +219,10 @@ RX_STATIC_TESTABLE volatile rx_tmr_channel_regs_t* internal_tmr_get_regs(rx_tmr_
 static uint32_t internal_tmr_unit_mstp_bit(rx_tmr_channel_t channel)
 {
   if ((channel == k_tmr_channel_0) || (channel == k_tmr_channel_1)) {
-    return (uint32_t)k_tmr_mstpcra_mstpa5;
+    return k_tmr_mstpcra_mstpa5;
   }
   if ((channel == k_tmr_channel_2) || (channel == k_tmr_channel_3)) {
-    return (uint32_t)k_tmr_mstpcra_mstpa4;
+    return k_tmr_mstpcra_mstpa4;
   }
   return 0U;
 }
@@ -326,28 +326,28 @@ static uint8_t internal_tmr_compose_tcr(const rx_tmr_config_t* config)
 {
   uint8_t tcr = 0U;
 
-  if ((config->irq_mask & (uint8_t)k_tmr_irq_cmp_match_a) != 0U) {
-    tcr |= (uint8_t)k_tmr_tcr_cmiea;
+  if ((config->irq_mask & k_tmr_irq_cmp_match_a) != 0U) {
+    tcr |= k_tmr_tcr_cmiea;
   }
-  if ((config->irq_mask & (uint8_t)k_tmr_irq_cmp_match_b) != 0U) {
-    tcr |= (uint8_t)k_tmr_tcr_cmieb;
+  if ((config->irq_mask & k_tmr_irq_cmp_match_b) != 0U) {
+    tcr |= k_tmr_tcr_cmieb;
   }
-  if ((config->irq_mask & (uint8_t)k_tmr_irq_overflow) != 0U) {
-    tcr |= (uint8_t)k_tmr_tcr_ovie;
+  if ((config->irq_mask & k_tmr_irq_overflow) != 0U) {
+    tcr |= k_tmr_tcr_ovie;
   }
 
   switch (config->counter_clear) {
     case k_tmr_clear_disabled:
-      tcr |= (uint8_t)k_tmr_tcr_cclr_disabled;
+      tcr |= k_tmr_tcr_cclr_disabled;
       break;
     case k_tmr_clear_cmp_match_a:
-      tcr |= (uint8_t)k_tmr_tcr_cclr_cmp_match_a;
+      tcr |= k_tmr_tcr_cclr_cmp_match_a;
       break;
     case k_tmr_clear_cmp_match_b:
-      tcr |= (uint8_t)k_tmr_tcr_cclr_cmp_match_b;
+      tcr |= k_tmr_tcr_cclr_cmp_match_b;
       break;
     case k_tmr_clear_external_reset:
-      tcr |= (uint8_t)k_tmr_tcr_cclr_external_sig;
+      tcr |= k_tmr_tcr_cclr_external_sig;
       break;
     default:
       /* Caller validates */
@@ -377,22 +377,22 @@ static uint8_t internal_tmr_compose_tcsr(const rx_tmr_config_t* config)
  */
 static rx_err_t internal_tmr_validate_config(const rx_tmr_config_t* config)
 {
-  if ((uint8_t)config->channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)config->channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   if ((config->mode != k_tmr_mode_8bit_independent) && (config->mode != k_tmr_mode_16bit_cascade)) {
     return k_rx_err_invalid_arg;
   }
-  if ((uint8_t)config->counter_clear > (uint8_t)k_tmr_clear_external_reset) {
+  if ((uint8_t)config->counter_clear > k_tmr_clear_external_reset) {
     return k_rx_err_invalid_arg;
   }
-  if ((uint8_t)config->output_a > (uint8_t)k_tmr_output_toggle) {
+  if ((uint8_t)config->output_a > k_tmr_output_toggle) {
     return k_rx_err_invalid_arg;
   }
-  if ((uint8_t)config->output_b > (uint8_t)k_tmr_output_toggle) {
+  if ((uint8_t)config->output_b > k_tmr_output_toggle) {
     return k_rx_err_invalid_arg;
   }
-  if ((uint8_t)config->clock_source > (uint8_t)k_tmr_clock_external_both) {
+  if ((uint8_t)config->clock_source > k_tmr_clock_external_both) {
     return k_rx_err_invalid_arg;
   }
   /* Cascade mode requires an even channel */
@@ -417,8 +417,8 @@ static void internal_tmr_program_channel(rx_tmr_channel_t channel, const rx_tmr_
   regs->tccr  = 0U; /* Stop counter before programming */
   regs->tcr   = internal_tmr_compose_tcr(config);
   regs->tcsr  = internal_tmr_compose_tcsr(config);
-  regs->tcora = (uint8_t)k_tmr_max_tcora;
-  regs->tcorb = (uint8_t)k_tmr_max_tcora;
+  regs->tcora = k_tmr_max_tcora;
+  regs->tcorb = k_tmr_max_tcora;
   regs->tcnt  = 0U;
 }
 
@@ -472,11 +472,11 @@ rx_err_t rx_tmr_init(const rx_tmr_config_t* config)
     internal_tmr_program_channel(odd, &odd_cfg);
 
     volatile rx_tmr_channel_regs_t* odd_regs = internal_tmr_get_regs(odd);
-    odd_regs->tccr                           = (uint8_t)k_tmr_tccr_css_cascade;
+    odd_regs->tccr                           = k_tmr_tccr_css_cascade;
 
     s_tmr_mode[(uint8_t)odd]         = k_tmr_mode_16bit_cascade;
     s_tmr_initialized[(uint8_t)odd]  = true;
-    s_tmr_clock_bits[(uint8_t)odd]   = (uint8_t)k_tmr_tccr_css_cascade;
+    s_tmr_clock_bits[(uint8_t)odd]   = k_tmr_tccr_css_cascade;
     s_tmr_clock_source[(uint8_t)odd] = config->clock_source;
   }
 
@@ -492,7 +492,7 @@ rx_err_t rx_tmr_init(const rx_tmr_config_t* config)
 
 rx_err_t rx_tmr_start(rx_tmr_channel_t channel)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   const uint8_t idx = (uint8_t)channel;
@@ -505,9 +505,9 @@ rx_err_t rx_tmr_start(rx_tmr_channel_t channel)
    * kept after a stop/start cycle. */
   if (s_tmr_mode[idx] == k_tmr_mode_16bit_cascade) {
     if (channel == k_tmr_channel_0) {
-      tmr1()->tccr = (uint8_t)k_tmr_tccr_css_cascade;
+      tmr1()->tccr = k_tmr_tccr_css_cascade;
     } else if (channel == k_tmr_channel_2) {
-      tmr3()->tccr = (uint8_t)k_tmr_tccr_css_cascade;
+      tmr3()->tccr = k_tmr_tccr_css_cascade;
     }
   }
   return k_rx_ok;
@@ -515,7 +515,7 @@ rx_err_t rx_tmr_start(rx_tmr_channel_t channel)
 
 rx_err_t rx_tmr_stop(rx_tmr_channel_t channel)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   const uint8_t idx = (uint8_t)channel;
@@ -537,7 +537,7 @@ rx_err_t rx_tmr_stop(rx_tmr_channel_t channel)
 rx_err_t rx_tmr_read(rx_tmr_channel_t channel, uint16_t* value)
 {
   RX_CHECK_NULL_PTR(value, s_tag, "Value pointer is NULL");
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   const uint8_t idx = (uint8_t)channel;
@@ -556,7 +556,7 @@ rx_err_t rx_tmr_read(rx_tmr_channel_t channel, uint16_t* value)
 
 rx_err_t rx_tmr_set_period_us(rx_tmr_channel_t channel, uint32_t period_us)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   const uint8_t idx = (uint8_t)channel;
@@ -582,8 +582,8 @@ rx_err_t rx_tmr_set_period_us(rx_tmr_channel_t channel, uint32_t period_us)
   const uint64_t compare = ticks - 1U;
 
   const uint32_t max_compare = (s_tmr_mode[idx] == k_tmr_mode_16bit_cascade)
-                                 ? (uint32_t)k_tmr_max_tcora_16bit
-                                 : (uint32_t)k_tmr_max_tcora;
+                                 ? k_tmr_max_tcora_16bit
+                                 : k_tmr_max_tcora;
   if (compare > (uint64_t)max_compare) {
     rx_log_error_val(s_tag, "Period too large for divider", period_us);
     return k_rx_err_invalid_arg;
@@ -601,7 +601,7 @@ rx_err_t rx_tmr_set_period_us(rx_tmr_channel_t channel, uint32_t period_us)
 
 rx_err_t rx_tmr_register_compare_match_isr(rx_tmr_channel_t channel, rx_tmr_isr_callback_t callback)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
   s_tmr_callback[(uint8_t)channel] = callback;
@@ -610,7 +610,7 @@ rx_err_t rx_tmr_register_compare_match_isr(rx_tmr_channel_t channel, rx_tmr_isr_
 
 rx_err_t rx_tmr_deinit(rx_tmr_channel_t channel)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return k_rx_err_invalid_arg;
   }
 
@@ -665,7 +665,7 @@ rx_err_t rx_tmr_deinit(rx_tmr_channel_t channel)
  */
 void rx_tmr_isr_dispatch(rx_tmr_channel_t channel)
 {
-  if ((uint8_t)channel >= (uint8_t)k_tmr_channel_count_internal) {
+  if ((uint8_t)channel >= k_tmr_channel_count_internal) {
     return;
   }
   rx_tmr_isr_callback_t cb = s_tmr_callback[(uint8_t)channel];
